Restructure project and add more validations for categories
