+++ src/main/java/com/cicerone/activity/question/QuestionType.java
@@ -1,4 +1,4 @@
-package com.cicerone.model;
+package com.cicerone.activity.question;
 
 public enum QuestionType {
     SINGLE_CHOICE, MULTIPLE_CHOICE, TRUE_OR_FALSE
