+++ src/main/java/com/cicerone/category/CategoryRepository.java
@@ -1,6 +1,5 @@
-package com.cicerone.repository;
+package com.cicerone.category;
 
-import com.cicerone.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
