+++ src/main/java/com/cicerone/course/Course.java
@@ -1,5 +1,6 @@
-package com.cicerone.model;
+package com.cicerone.course;
 
+import com.cicerone.subcategory.Subcategory;
 import com.cicerone.util.validation.Validator;
 
 import javax.persistence.*;
