+++ src/main/java/com/cicerone/section/Section.java
@@ -1,11 +1,12 @@
-package com.cicerone.model;
+package com.cicerone.section;
 
+import com.cicerone.course.Course;
 import com.cicerone.util.validation.Validator;
 
 import javax.persistence.*;
 
-@MappedSuperclass
-public abstract class Activity {
+@Entity
+public class Section {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -13,26 +14,27 @@ public abstract class Activity {
     private String title;
     private String code;
     private boolean disabled;
+    private boolean exam;
 
     @Column(name = "orderPosition")
     private Integer order;
 
     @ManyToOne
-    @JoinColumn(name = "sectionId")
-    private Section section;
+    @JoinColumn(name = "courseId")
+    private Course course;
 
     @Deprecated
-    public Activity(){};
+    public Section(){};
 
     // Required fields only
-    public Activity(String title, String code, Section section) {
+    public Section(String title, String code, Course course) {
         Validator.isNotBlankString(title);
         Validator.isValidCode(code);
-        Validator.isNotNullObject(section);
+        Validator.isNotNullObject(course);
 
         this.title = title;
         this.code = code;
-        this.section = section;
+        this.course = course;
         this.disabled = true;
     }
