+++ src/main/java/com/cicerone/util/io/HTMLGenerator.java
@@ -1,8 +1,12 @@
 package com.cicerone.util.io;
 
-import com.cicerone.model.*;
+import com.cicerone.category.Category;
+import com.cicerone.course.Course;
+import com.cicerone.subcategory.Subcategory;
 
-import java.io.*;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
