+++ src/main/resources/ValidationMessages.properties
@@ -1,4 +1,6 @@
 # Category
-category.title.notblank = O nome da categoria é obrigatório
-category.code.notblank = O código da categoria é obrigatório
+category.title.not.blank = O nome da categoria é obrigatório
+category.title.size = O nome da categoria deve ter ao menos 3 caracteres
+category.code.not.blank = O código da categoria é obrigatório
 category.code.pattern = O código deve conter somente letras minúsculas e hífens
+category.code.positive = A ordem deve ser maior que 0
