+++ src/test/java/com/cicerone/course/CourseTest.java
@@ -1,5 +1,7 @@
-package com.cicerone.model;
+package com.cicerone.course;
 
+import com.cicerone.category.Category;
+import com.cicerone.subcategory.Subcategory;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
