+++ src/test/java/com/cicerone/db/dao/SubcategoryDAOTest.java
@@ -1,7 +1,7 @@
 package com.cicerone.db.dao;
 
-import com.cicerone.model.Category;
-import com.cicerone.model.Subcategory;
+import com.cicerone.category.Category;
+import com.cicerone.subcategory.Subcategory;
 import com.cicerone.util.builder.CategoryBuilder;
 import com.cicerone.util.builder.SubcategoryBuilder;
 import com.cicerone.util.db.JPAUtil;
